Propose a new schema and versioning scheme for ONS version 1.0.0
This is part 1 of a 3 part proposal to make the ONS spec modular,
extentible, and forwards and backwards compatible.

The new spec would use three part semantic versioning.

The version schema proposed here is the lowest level of the spec.

A structure for storing sections, which are arrays of objects is
proposed.

The list of section names is not defined here, that will be the
job of the 1.0 spec, and the 1.0.0 spec will define the supported
tags and types of objects.

New sections can be added to the spec by incrementing the minor
version number (1.0 -> 1.1) and new tags and types can be added
by incrementing the patch level (1.0.0 -> 1.0.1).

The other two parts of the proposal will be presented as seperate
 pull requests which build on this one.