Update multi-player-storage.md
As a js/react beginner, I had to refer to the video in order to find the right place to add the isLocal() condition code.  Wording around that was vague for me.  Not sure this is much better, but something to consider.   Images are broken in md but work for the main website.  Changed references from index.jsx to index.js, per the code generated.  Also note that people coming to this will likely not have an id yet